Moved widgets into folders to structure the home widget folder
