+++ lib/widgets/home/searching/launch_search_delegate.dart
@@ -3,7 +3,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:space_x_tracker/providers/launch_provider.dart';
 import 'package:space_x_tracker/providers/models/launch.dart';
-import 'package:space_x_tracker/widgets/home/card_list_view.dart';
+import 'package:space_x_tracker/widgets/home/list/card_list_view.dart';
 import 'package:space_x_tracker/widgets/home/no_launch_results_message.dart';
 import 'package:space_x_tracker/widgets/home/searching/no_search_results_message.dart';
